feat: Move ConnectionRepository to new folder structure
