+++ src/script/entity/Conversation.test.ts
@@ -23,7 +23,7 @@ import {ConnectionStatus} from '@wireapp/api-client/lib/connection/';
 import {CONVERSATION_TYPE} from '@wireapp/api-client/lib/conversation/';
 
 import {ClientEntity} from 'Repositories/client/ClientEntity';
-import {ConnectionMapper} from 'src/script/connection/ConnectionMapper';
+import {ConnectionMapper} from 'Repositories/connection/ConnectionMapper';
 import {ConversationMapper} from 'src/script/conversation/ConversationMapper';
 import {NOTIFICATION_STATE} from 'src/script/conversation/NotificationSetting';
 import 'src/script/localization/Localizer';
