+++ src/script/entity/User/User.ts
@@ -28,13 +28,13 @@ import {WebAppEvents} from '@wireapp/webapp-events';
 
 import type {AssetRemoteData} from 'Repositories/assets/AssetRemoteData';
 import type {ClientEntity} from 'Repositories/client/ClientEntity';
+import {ConnectionEntity} from 'Repositories/connection/ConnectionEntity';
 import {t} from 'Util/LocalizerUtil';
 import {clamp} from 'Util/NumberUtil';
 import {getFirstChar} from 'Util/StringUtil';
 import {TIME_IN_MILLIS} from 'Util/TimeUtil';
 
 import {ACCENT_ID} from '../../Config';
-import {ConnectionEntity} from '../../connection/ConnectionEntity';
 import {ROLE as TEAM_ROLE} from '../../user/UserPermission';
 
 export class User {
