+++ src/script/main/app.ts
@@ -40,6 +40,8 @@ import {CacheRepository} from 'Repositories/cache/CacheRepository';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
 import {ClientRepository, ClientService} from 'Repositories/client';
 import {getClientMLSConfig} from 'Repositories/client/clientMLSConfig';
+import {ConnectionRepository} from 'Repositories/connection/ConnectionRepository';
+import {ConnectionService} from 'Repositories/connection/ConnectionService';
 import {initializeDataDog} from 'Util/DataDog';
 import {DebugUtil} from 'Util/DebugUtil';
 import {Environment} from 'Util/Environment';
@@ -54,8 +56,6 @@ import '../../style/default.less';
 import {SIGN_OUT_REASON} from '../auth/SignOutReason';
 import {URLParameter} from '../auth/URLParameter';
 import {Configuration} from '../Config';
-import {ConnectionRepository} from '../connection/ConnectionRepository';
-import {ConnectionService} from '../connection/ConnectionService';
 import {ConversationRepository} from '../conversation/ConversationRepository';
 import {ConversationService} from '../conversation/ConversationService';
 import {ConversationVerificationState} from '../conversation/ConversationVerificationState';
