+++ src/script/notification/NotificationRepository.ts
@@ -33,6 +33,7 @@ import {AudioType} from 'Repositories/audio/AudioType';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
 import {CallingViewMode, CallState} from 'Repositories/calling/CallState';
 import {TERMINATION_REASON} from 'Repositories/calling/enum/TerminationReason';
+import type {ConnectionEntity} from 'Repositories/connection/ConnectionEntity';
 import {Declension, t, getUserName} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
 import {getRenderedTextContent} from 'Util/messageRenderer';
@@ -42,7 +43,6 @@ import {ValidationUtilError} from 'Util/ValidationUtil';
 
 import {PermissionState} from './PermissionState';
 
-import type {ConnectionEntity} from '../connection/ConnectionEntity';
 import {ConversationEphemeralHandler} from '../conversation/ConversationEphemeralHandler';
 import type {ConversationRepository} from '../conversation/ConversationRepository';
 import {ConversationState} from '../conversation/ConversationState';
