+++ src/script/page/RightSidebar/ConversationDetails/ConversationDetails.test.tsx
@@ -20,13 +20,13 @@
 import {act, render} from '@testing-library/react';
 import {CONVERSATION_TYPE} from '@wireapp/api-client/lib/conversation';
 
+import {ConnectionRepository} from 'Repositories/connection/ConnectionRepository';
 import {Conversation} from 'src/script/entity/Conversation';
 import {createUuid} from 'Util/uuid';
 
 import {ConversationDetails} from './ConversationDetails';
 
 import {TestFactory} from '../../../../../test/helper/TestFactory';
-import {ConnectionRepository} from '../../../connection/ConnectionRepository';
 import {ConversationRepository} from '../../../conversation/ConversationRepository';
 import {ConversationRoleRepository} from '../../../conversation/ConversationRoleRepository';
 import {MessageRepository} from '../../../conversation/MessageRepository';
