+++ src/script/repositories/connection/ConnectionEntity.ts
