+++ src/script/repositories/connection/ConnectionState.ts
