+++ src/script/user/UserRepository.test.ts
@@ -31,6 +31,7 @@ import {WebAppEvents} from '@wireapp/webapp-events';
 import {AssetRepository} from 'Repositories/assets/AssetRepository';
 import {ClientRepository} from 'Repositories/client';
 import {ClientMapper} from 'Repositories/client/ClientMapper';
+import {ConnectionEntity} from 'Repositories/connection/ConnectionEntity';
 import {entities} from 'test/api/payloads';
 import {TestFactory} from 'test/helper/TestFactory';
 import {generateAPIUser} from 'test/helper/UserGenerator';
@@ -41,7 +42,6 @@ import {UserRepository} from './UserRepository';
 import {UserService} from './UserService';
 import {UserState} from './UserState';
 
-import {ConnectionEntity} from '../connection/ConnectionEntity';
 import {User} from '../entity/User';
 import {EventRepository} from '../event/EventRepository';
 import {PropertiesRepository} from '../properties/PropertiesRepository';
