+++ src/script/view_model/MainViewModel.ts
@@ -24,13 +24,13 @@ import type {AudioRepository} from 'Repositories/audio/AudioRepository';
 import type {BackupRepository} from 'Repositories/backup/BackupRepository';
 import type {CallingRepository} from 'Repositories/calling/CallingRepository';
 import type {ClientRepository} from 'Repositories/client';
+import type {ConnectionRepository} from 'Repositories/connection/ConnectionRepository';
 
 import {ActionsViewModel} from './ActionsViewModel';
 import {CallingViewModel} from './CallingViewModel';
 import {ContentViewModel} from './ContentViewModel';
 import {ListViewModel} from './ListViewModel';
 
-import type {ConnectionRepository} from '../connection/ConnectionRepository';
 import type {ConversationRepository} from '../conversation/ConversationRepository';
 import type {MessageRepository} from '../conversation/MessageRepository';
 import type {CryptographyRepository} from '../cryptography/CryptographyRepository';
