+++ test/helper/TestFactory.js
@@ -35,8 +35,8 @@ import {ClientEntity} from 'Repositories/client/ClientEntity';
 import {ClientRepository} from 'Repositories/client/ClientRepository';
 import {ClientService} from 'Repositories/client/ClientService';
 import {ClientState} from 'Repositories/client/ClientState';
-import {ConnectionRepository} from 'src/script/connection/ConnectionRepository';
-import {ConnectionService} from 'src/script/connection/ConnectionService';
+import {ConnectionRepository} from 'Repositories/connection/ConnectionRepository';
+import {ConnectionService} from 'Repositories/connection/ConnectionService';
 import {ConversationRepository} from 'src/script/conversation/ConversationRepository';
 import {ConversationService} from 'src/script/conversation/ConversationService';
 import {ConversationState} from 'src/script/conversation/ConversationState';
